Adds MarkdownElement to the rendering framework
MarkdownElements are like TextElements, but with their contents
rendered as Markdown. You can create an inline Markdown element
with the `formatted` built-in.

Additionally, if you provide a Frame when you create a MarkdownElement,
you can add elements like `[Text](:procedure)`, which will create a
button that calls `(procedure)` in the provided environment when clicked.

The motd is now logged as a MarkdownElement and now includes highlighted
syntax and buttons to launch demos of the various listed features.

This also makes some miscellaneous fixes, mostly to styles and rendering.