Adds SchemeList wrapper type
Pairs are no longer directly iterable. Instead, whenever we want to
guarantee the a PairOrEmpty is a Scheme list, we wrap it in
SchemeList, which then allows iteration.

We then use this new wrapper to enforce stricter typing of Scheme
lists throughout the interpreter. The builder supports auto-wrapping
and unwrapping.