Major refactor of data types (#76)
* Refactor base Scheme type to be Value

Previously, everything that could be touched by Scheme code was
an Expression. However, many Expression subtypes can't actually
be evaluated.

This refactor changes the base Scheme type to be a Value, with a
limited subset of evaluate-able types remaining Expressions (which
itself inherits from Value).

* Adds SchemeList wrapper type

Pairs are no longer directly iterable. Instead, whenever we want to
guarantee the a PairOrEmpty is a Scheme list, we wrap it in
SchemeList, which then allows iteration.

We then use this new wrapper to enforce stricter typing of Scheme
lists throughout the interpreter. The builder supports auto-wrapping
and unwrapping.

* Update private impl in Travis config